fix: section header is displayed even when page not found
