+++ apps/ai-dial-admin/src/components/Breadcrumbs/models.ts
@@ -1,0 +1,15 @@
+import { MenuI18nKey } from '@/src/constants/i18n';
+
+export interface BreadcrumbConfig {
+  segments: {
+    name: string;
+    i18nKey?: MenuI18nKey;
+    href?: boolean;
+  }[];
+}
+
+export interface Breadcrumb {
+  key?: MenuI18nKey;
+  name: string;
+  href: string;
+}
